+++ packages/nextjs-analytics/package.json
@@ -16,7 +16,6 @@
   ],
   "dependencies": {
     "@ogcio/building-blocks-sdk": "^0.1.12",
-    "next": "14.2.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
