+++ packages/nextjs-analytics/src/types.ts
@@ -1,32 +1,16 @@
-import type { getBuildingBlockSDK } from "@ogcio/building-blocks-sdk";
+import type {
+  BuildingBlocksSDK,
+  getBuildingBlockSDK,
+} from "@ogcio/building-blocks-sdk";
 
 type AnalyticsClientProps = ReturnType<typeof getBuildingBlockSDK>["analytics"];
 
 type AnalyticsTrackerProps = Parameters<
   AnalyticsClientProps["setTrackingContext"]
->[0];
+>[0] & { pathname?: string };
 
-interface TrackEventProps {
-  event: {
-    category: string;
-    action: string;
-    name?: string;
-    value?: number;
-  };
-  metadataOverride?: {
-    url?: string;
-    userAgent?: string;
-    referrer?: string;
-    language?: string;
-    screenResolution?: string;
-  };
-  contextOverride?: {
-    userId?: string;
-    customDimensions?: {
-      [key: string]: string;
-    };
-  };
-  siteIds?: number[];
-}
+type TrackEventProps = Parameters<
+  BuildingBlocksSDK["analytics"]["track"]["event"]
+>[0];
 
 export type { AnalyticsClientProps, AnalyticsTrackerProps, TrackEventProps };
