+++ packages/nextjs-analytics/tsconfig.json
@@ -5,7 +5,9 @@
     "types": ["./src/custom.d.ts"],
     "module": "ESNext",
     "moduleResolution": "bundler",
+    "composite": true,
     "strictNullChecks": true,
+    "jsx": "preserve",
     "outDir": "dist",
     "rootDir": "src",
   },
